+++ src/main/java/hls/analyzer/EXTXTARGETDURATION.java
@@ -14,14 +14,14 @@ public class EXTXTARGETDURATION extends Validator{
 	@Override
 	public List<ValidationReport> isValid() throws IOException {
 		List<ValidationReport> errorMsgs = new ArrayList<ValidationReport>(); 
-		int duration = -1;
+		Float duration = (float) -1;
 		for (String dataItem : _dataFileArray) {
 			int lineNumber = _dataFileArray.indexOf(dataItem);
 			if(!dataItem.isEmpty() && dataItem.matches(Constants.durationRegex)){
 				if(duration < 0){
 					String durationValue = UtilHelper.parseStringAttr(dataItem, Constants.durationRegex);
 					if(durationValue.matches(Constants.intRegex)){
-						duration =  Integer.parseInt(UtilHelper.parseStringAttr(dataItem, Constants.durationRegex));
+						duration =  Float.parseFloat(UtilHelper.parseStringAttr(dataItem, Constants.durationRegex));
 						errorMsgs.addAll(new EXTINF(_dataFileArray,_fileName, duration).isValid());
 					}else{
 						errorMsgs.add(new ValidationReport(lineNumber,Constants.EXTXTARGETDURATION,_fileName,"EXT-X-TARGETDURATION duration number should be an integer or float"));
@@ -30,7 +30,7 @@ public class EXTXTARGETDURATION extends Validator{
 			}
 		}
 		
-		if(duration <0){
+		if(duration < 0){
 			errorMsgs.add(new ValidationReport(null,Constants.EXTXTARGETDURATION,_fileName,"EXT-X-TARGETDURATION tag not detected which make this file invalid."));
 		}
